ai-layer/requirements.txt
fastapi
uvicorn
torch
torchvision
scikit-learn
firebase-admin
numpy
pillow
requests
python-multipart
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
nltk